Move internals into /src folder, format src as a package
